Fix conversions for ThreadedHistogram initialization

+++ src/core/module/ModuleManager.cpp
@@ -623,7 +623,7 @@ void ModuleManager::run(RandomNumberGenerator& seeder) {
     auto plot = global_config.get<bool>("performance_plots");
 
     // Default to no additional thread without multithreading
-    auto threads_num = global_config.get<size_t>("workers");
+    auto threads_num = global_config.get<unsigned int>("workers");
     size_t max_buffer_size = 1;
 
     // See if we can run in parallel with how many workers
@@ -653,8 +653,11 @@ void ModuleManager::run(RandomNumberGenerator& seeder) {
 
     // Book performance histograms
     if(global_config.get<bool>("performance_plots")) {
-        buffer_fill_level_ = CreateHistogram<TH1D>(
-            "buffer_fill_level", "Buffer fill level;# buffered events;# events", max_buffer_size, 0, max_buffer_size);
+        buffer_fill_level_ = CreateHistogram<TH1D>("buffer_fill_level",
+                                                   "Buffer fill level;# buffered events;# events",
+                                                   static_cast<int>(max_buffer_size),
+                                                   0,
+                                                   static_cast<double>(max_buffer_size));
         event_time_ = CreateHistogram<TH1D>("event_time", "processing time per event;time [s];# events", 1000, 0, 10);
         for(auto& module : modules_) {
             auto identifier = module->get_identifier().getIdentifier();
